Added nickname detection and other minor tweaks
Added the ability to have a table of nicknames and have the chat parser search the messages for instances of that nickname.

Adjusted some data formats so dates are now datetime and can be properly indexed.

Stat tracking should now show total of UNIQUE users to process and not simply the number of every message.

There's some functionality for populating a database based on what chat JSONs have been already downloaded, but I ended up not utilizing it for anything so it might not work as it's not fully implemented and tested.